refactor: move current code into 'legacy' folder behind `Backend` interface
