+++ src/ExtensionBackend.ts
@@ -1,0 +1,5 @@
+import * as vscode from "vscode";
+
+export interface ExtensionBackend extends vscode.Disposable {
+  reInitialize(): Promise<void>;
+}
