+++ src/legacy/ConfigJsonSchemaProvider.ts
@@ -1,6 +1,6 @@
 import * as vscode from "vscode";
-import { Logger } from "./logger";
-import { RacyCacheTextDownloader, TextDownloader } from "./TextDownloader";
+import { Logger } from "../logger";
+import { RacyCacheTextDownloader, TextDownloader } from "../utils";
 import { FolderInfos } from "./WorkspaceService";
 
 /** Provides the dprint configuration JSON schema to vscode. */
