+++ src/utils/TextDownloader.ts
